# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Nuvana Production API
  description: |
    Complete API documentation for the Nuvana Education Platform backend.

    **Authentication**: All endpoints (except those marked Public) require a Bearer JWT token in the `Authorization` header.
    The `x-school-id` header is required for tenant isolation on most endpoints.

    **Roles**: super_admin, school_admin, teacher, student

    **Rate Limiting**: Global 100 req/min. AI endpoints: 5 req/min.
  version: 1.0.0
  contact:
    name: Nuvana Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.nuvana.in
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterSuperAdminDto:
      type: object
      required: [email, password, name, secret]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        secret:
          type: string
    RegisterUserDto:
      type: object
      required: [email, name, role_id, temporaryPassword]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role_id:
          type: integer
          description: "1=super_admin, 2=school_admin, 3=teacher, 4=student"
        school_id:
          type: string
          format: uuid
        temporaryPassword:
          type: string
          minLength: 8
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        school_id:
          type: string
          format: uuid
        expectedRole:
          type: string
    ResetPasswordDto:
      type: object
      required: [userId, newPassword]
      properties:
        userId:
          type: string
          format: uuid
        newPassword:
          type: string
          minLength: 8
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            name:
              type: string
            role:
              type: string
            school_id:
              type: string
              format: uuid

    # ─── Users ──────────────────────────────────────────
...
    WhatsAppSendDto:
      type: object
      required: [useCase, data, senderId, schoolId]
      properties:
        useCase:
          type: string
          enum: [school_announcement]
        data:
          type: object
          description: Template-specific data
          properties:
            parentName:
              type: string
            studentName:
              type: string
            messageText:
              type: string
            schoolName:
              type: string
        phoneNumber:
          type: string
        recipients:
          type: array
          items:
            type: object
            properties:
              phoneNumber:
                type: string
              recipientId:
                type: string
        senderId:
          type: string
        schoolId:
          type: string

    # ─── Common ─────────────────────────────────────────
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
security:
  - BearerAuth: []
paths:
...
  /auth/super-admin/register:
    post:
      tags: [Auth]
      summary: Register super admin (Public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSuperAdminDto'
      responses:
        '201':
          description: Super admin registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user (super_admin, school_admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  name:
                    type: string
  /auth/login:
    post:
      tags: [Auth]
      summary: Login (Public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password (Public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token (Public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
...
  /whatsapp/send:
    post:
      tags: [WhatsApp]
      summary: Send WhatsApp message (queued)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhatsAppSendDto'
      responses:
        '201':
          description: Message queued
          content:
            application/json:
              schema:
                type: object
  /whatsapp/recipients/class/{classId}:
    get:
      tags: [WhatsApp]
      summary: Get WhatsApp recipients for class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipients list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    phoneNumber:
                      type: string
                    studentName:
                      type: string
                    parentName:
                      type: string
  /whatsapp/recipients/student/{studentId}:
    get:
      tags: [WhatsApp]
      summary: Get WhatsApp recipient for student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipient data
          content:
            application/json:
              schema:
                type: object
  /whatsapp/history:
    get:
      tags: [WhatsApp]
      summary: Get WhatsApp message history
      parameters:
        - name: schoolId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object